requirements.txt
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
scikit-learn>=0.24.0
tqdm>=4.62.0
Pillow>=8.3.0
torch>=1.9.0
torchvision>=0.10.0
